refactor: (#164) 도메인 분리 및 서비스 이관으로 인한 테스트코드 패키지를 변경한다
