+++ src/test/java/spring/backend/activity/application/QuickStartActivitySelectServiceTest.java
@@ -8,15 +8,16 @@ import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
 import spring.backend.activity.domain.entity.Activity;
-import spring.backend.activity.domain.entity.QuickStart;
+import spring.backend.quickstart.domain.entity.QuickStart;
 import spring.backend.activity.domain.repository.ActivityRepository;
-import spring.backend.activity.domain.repository.QuickStartRepository;
+import spring.backend.quickstart.domain.repository.QuickStartRepository;
+import spring.backend.activity.domain.service.FinishActivityAutoService;
 import spring.backend.activity.domain.value.Keyword;
 import spring.backend.activity.domain.value.Type;
 import spring.backend.activity.dto.request.QuickStartActivitySelectRequest;
 import spring.backend.activity.dto.response.QuickStartActivitySelectResponse;
 import spring.backend.activity.exception.ActivityErrorCode;
-import spring.backend.activity.exception.QuickStartErrorCode;
+import spring.backend.quickstart.exception.QuickStartErrorCode;
 import spring.backend.core.exception.DomainException;
 import spring.backend.member.domain.entity.Member;
 import spring.backend.member.domain.value.Role;
